fix: exception management improved
