+++ srcs/server/src/friendship/exceptions/index.ts
@@ -1,4 +1,4 @@
-export * from "./friendshipAlreadyExists.exception";
-export * from "./friendshipNotFound.exception";
-export * from "./friendshipRejected.exception";
-export * from "./friendshipRequested.exception";
+export * from './friendshipAlreadyExists.exception';
+export * from './friendshipNotFound.exception';
+export * from './friendshipRejected.exception';
+export * from './friendshipRequested.exception';
