+++ srcs/server/src/match/exceptions/index.ts
@@ -1,0 +1,5 @@
+export * from './matchAlreadyExits.exception';
+export * from './matchNotFound.exception';
+export * from './matchWithOnePlayer.exception';
+export * from './noMatchesInDatabase.exception';
+export * from './playersNotAvailable.exception';
