+++ srcs/server/src/match/exceptions/matchNotFound.exception.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class MatchNotFoundException extends HttpException {
+  constructor(id: number) {
+    super(`Match #${id} not found`, HttpStatus.NOT_FOUND);
+  }
+}
