+++ srcs/server/src/match/exceptions/noMatchesInDatabase.exception.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class NoMatchesInDatabaseException extends HttpException {
+  constructor() {
+    super(`No matches in database`, HttpStatus.NO_CONTENT);
+  }
+}
