+++ srcs/server/src/match/exceptions/playersNotAvailable.exception.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class PlayersNotAvailableException extends HttpException {
+  constructor() {
+    super(`At least one of the player is not available to play`, HttpStatus.OK);
+  }
+}
