+++ srcs/server/src/user/exceptions/index.ts
@@ -1,0 +1,3 @@
+export * from './noUsersInDatabase.exception';
+export * from './userAlreadyExists.exception';
+export * from './userNotFound.exception';
