+++ srcs/server/src/user/exceptions/noUsersInDatabase.exception.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class NoUsersInDatabaseException extends HttpException {
+  constructor() {
+    super(`No users in database`, HttpStatus.NO_CONTENT);
+  }
+}
