+++ srcs/server/src/user/exceptions/userAlreadyExists.exception.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class UserAlreadyExistsException extends HttpException {
+  constructor(username: string) {
+    super(`User "${username}" already exists`, HttpStatus.CONFLICT);
+  }
+}
