+++ srcs/server/src/user/exceptions/userNotFound.exception.ts
@@ -1,0 +1,7 @@
+import { HttpException, HttpStatus } from '@nestjs/common';
+
+export class UserNotFoundException extends HttpException {
+  constructor(id: number) {
+    super(`User #${id} not found`, HttpStatus.NOT_FOUND);
+  }
+}
